Struct indextree::NodeId

pub struct NodeId { /* fields omitted */ }

A node identifier within a particular Arena.

This ID is used to get Node references from an Arena.

Methods

impl NodeId

Important traits for Ancestors<'a, T>
impl<'a, T> Iterator for Ancestors<'a, T> type Item = NodeId;

pub fn ancestors<T>(self, arena: &Arena<T>) -> Ancestors<T>

Returns an iterator of references to this node and its ancestors.

Use .skip(1) or call .next() once on the iterator to skip the node itself.

Examples

// arena
// `-- 1
//     |-- 1_1
//     |   `-- 1_1_1
//     |       `-- 1_1_1_1
//     _-- 1_2
//     `-- 1_3

let mut iter = n1_1_1.ancestors(&arena);
assert_eq!(iter.next(), Some(n1_1_1));
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), Some(n1));
assert_eq!(iter.next(), None);

Important traits for PrecedingSiblings<'a, T>
impl<'a, T> Iterator for PrecedingSiblings<'a, T> type Item = NodeId;

pub fn preceding_siblings<T>(self, arena: &Arena<T>) -> PrecedingSiblings<T>

Returns an iterator of references to this node and the siblings before it.

Use .skip(1) or call .next() once on the iterator to skip the node itself.

Examples

// arena
// `-- 1
//     |-- 1_1
//     |   `-- 1_1_1
//     |-- 1_2
//     `-- 1_3

let mut iter = n1_2.preceding_siblings(&arena);
assert_eq!(iter.next(), Some(n1_2));
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), None);

Important traits for FollowingSiblings<'a, T>
impl<'a, T> Iterator for FollowingSiblings<'a, T> type Item = NodeId;

pub fn following_siblings<T>(self, arena: &Arena<T>) -> FollowingSiblings<T>

Returns an iterator of references to this node and the siblings after it.

Use .skip(1) or call .next() once on the iterator to skip the node itself.

Examples

// arena
// `-- 1
//     |-- 1_1
//     |   `-- 1_1_1
//     |-- 1_2
//     `-- 1_3

let mut iter = n1_2.following_siblings(&arena);
assert_eq!(iter.next(), Some(n1_2));
assert_eq!(iter.next(), Some(n1_3));
assert_eq!(iter.next(), None);

Important traits for Children<'a, T>
impl<'a, T> Iterator for Children<'a, T> type Item = NodeId;

pub fn children<T>(self, arena: &Arena<T>) -> Children<T>

Returns an iterator of references to this node’s children.

Examples

// arena
// `-- 1
//     |-- 1_1
//     |   `-- 1_1_1
//     |-- 1_2
//     `-- 1_3

let mut iter = n1.children(&arena);
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), Some(n1_2));
assert_eq!(iter.next(), Some(n1_3));
assert_eq!(iter.next(), None);

Important traits for ReverseChildren<'a, T>
impl<'a, T> Iterator for ReverseChildren<'a, T> type Item = NodeId;

pub fn reverse_children<T>(self, arena: &Arena<T>) -> ReverseChildren<T>

Returns an iterator of references to this node’s children, in reverse order.

Examples

// arena
// `-- 1
//     |-- 1_1
//     |   `-- 1_1_1
//     |-- 1_2
//     `-- 1_3

let mut iter = n1.reverse_children(&arena);
assert_eq!(iter.next(), Some(n1_3));
assert_eq!(iter.next(), Some(n1_2));
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), None);

Important traits for Descendants<'a, T>
impl<'a, T> Iterator for Descendants<'a, T> type Item = NodeId;

pub fn descendants<T>(self, arena: &Arena<T>) -> Descendants<T>

Returns an iterator of references to this node and its descendants, in tree order.

Parent nodes appear before the descendants. Use .skip(1) or call .next() once on the iterator to skip the node itself.

Examples

// arena
// `-- 1
//     |-- 1_1
//     |   `-- 1_1_1
//     |       `-- 1_1_1_1
//     |-- 1_2
//     `-- 1_3

let mut iter = n1.descendants(&arena);
assert_eq!(iter.next(), Some(n1));
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), Some(n1_1_1));
assert_eq!(iter.next(), Some(n1_1_1_1));
assert_eq!(iter.next(), Some(n1_2));
assert_eq!(iter.next(), Some(n1_3));
assert_eq!(iter.next(), None);

Important traits for Traverse<'a, T>
impl<'a, T> Iterator for Traverse<'a, T> type Item = NodeEdge;

pub fn traverse<T>(self, arena: &Arena<T>) -> Traverse<T>

Returns an iterator of references to this node and its descendants, in tree order.

Examples

// arena
// `-- 1
//     |-- 1_1
//     |   `-- 1_1_1
//     |-- 1_2
//     `-- 1_3

let mut iter = n1.traverse(&arena);
assert_eq!(iter.next(), Some(NodeEdge::Start(n1)));
assert_eq!(iter.next(), Some(NodeEdge::Start(n1_1)));
assert_eq!(iter.next(), Some(NodeEdge::Start(n1_1_1)));
assert_eq!(iter.next(), Some(NodeEdge::End(n1_1_1)));
assert_eq!(iter.next(), Some(NodeEdge::End(n1_1)));
assert_eq!(iter.next(), Some(NodeEdge::Start(n1_2)));
assert_eq!(iter.next(), Some(NodeEdge::End(n1_2)));
assert_eq!(iter.next(), Some(NodeEdge::Start(n1_3)));
assert_eq!(iter.next(), Some(NodeEdge::End(n1_3)));
assert_eq!(iter.next(), Some(NodeEdge::End(n1)));
assert_eq!(iter.next(), None);

Important traits for ReverseTraverse<'a, T>
impl<'a, T> Iterator for ReverseTraverse<'a, T> type Item = NodeEdge;

pub fn reverse_traverse<T>(self, arena: &Arena<T>) -> ReverseTraverse<T>

Returns an iterator of references to this node and its descendants, in reverse tree order.

Examples

// arena
// `-- 1
//     |-- 1_1
//     |   `-- 1_1_1
//     |-- 1_2
//     `-- 1_3

let mut iter = n1.reverse_traverse(&arena);
assert_eq!(iter.next(), Some(NodeEdge::End(n1)));
assert_eq!(iter.next(), Some(NodeEdge::End(n1_3)));
assert_eq!(iter.next(), Some(NodeEdge::Start(n1_3)));
assert_eq!(iter.next(), Some(NodeEdge::End(n1_2)));
assert_eq!(iter.next(), Some(NodeEdge::Start(n1_2)));
assert_eq!(iter.next(), Some(NodeEdge::End(n1_1)));
assert_eq!(iter.next(), Some(NodeEdge::End(n1_1_1)));
assert_eq!(iter.next(), Some(NodeEdge::Start(n1_1_1)));
assert_eq!(iter.next(), Some(NodeEdge::Start(n1_1)));
assert_eq!(iter.next(), Some(NodeEdge::Start(n1)));
assert_eq!(iter.next(), None);

let traverse = n1.traverse(&arena).collect::<Vec<_>>();
let mut reverse = n1.reverse_traverse(&arena).collect::<Vec<_>>();
reverse.reverse();
assert_eq!(traverse, reverse);

pub fn detach<T>(self, arena: &mut Arena<T>)

Detaches a node from its parent and siblings. Children are not affected.

Examples

// arena
// `-- (implicit)
//     `-- 1
//         |-- 1_1
//         |   `-- 1_1_1
//         |-- 1_2 *
//         `-- 1_3

n1_2.detach(&mut arena);
// arena
// |-- (implicit)
// |   `-- 1
// |       |-- 1_1
// |       |   `-- 1_1_1
// |       `-- 1_3
// `-- (implicit)
//     `-- 1_2

assert!(arena[n1_2].parent().is_none());
assert!(arena[n1_2].previous_sibling().is_none());
assert!(arena[n1_2].next_sibling().is_none());

let mut iter = n1.descendants(&arena);
assert_eq!(iter.next(), Some(n1));
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), Some(n1_1_1));
assert_eq!(iter.next(), Some(n1_3));
assert_eq!(iter.next(), None);

pub fn append<T>(self, new_child: NodeId, arena: &mut Arena<T>)

Appends a new child to this node, after existing children.

Panics

Panics if:

the given new child is self, or

the current node or the given new child was already removed.

To check if the node is removed or not, use Node::is_removed().

Examples

let mut arena = Arena::new();
let n1 = arena.new_node("1");
let n1_1 = arena.new_node("1_1");
n1.append(n1_1, &mut arena);
let n1_2 = arena.new_node("1_2");
n1.append(n1_2, &mut arena);
let n1_3 = arena.new_node("1_3");
n1.append(n1_3, &mut arena);

// arena
// `-- 1
//     |-- 1_1
//     |-- 1_2
//     `-- 1_3

let mut iter = n1.descendants(&arena);
assert_eq!(iter.next(), Some(n1));
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), Some(n1_2));
assert_eq!(iter.next(), Some(n1_3));
assert_eq!(iter.next(), None);

pub fn checked_append<T>(
self,
new_child: NodeId,
arena: &mut Arena<T>
) -> Result<(), NodeError>

Appends a new child to this node, after existing children.

Failures

Returns NodeError::AppendSelf error if the given new child is self.

Returns NodeError::Removed error if the given new child or self is removed.

To check if the node is removed or not, use Node::is_removed().

Examples

let mut arena = Arena::new();
let n1 = arena.new_node("1");
assert!(n1.checked_append(n1, &mut arena).is_err());

let n1_1 = arena.new_node("1_1");
assert!(n1.checked_append(n1_1, &mut arena).is_ok());

pub fn prepend<T>(self, new_child: NodeId, arena: &mut Arena<T>)

Prepends a new child to this node, before existing children.

Panics

Panics if:

the given new child is self, or

the current node or the given new child was already removed.

To check if the node is removed or not, use Node::is_removed().

Examples

let mut arena = Arena::new();
let n1 = arena.new_node("1");
let n1_1 = arena.new_node("1_1");
n1.prepend(n1_1, &mut arena);
let n1_2 = arena.new_node("1_2");
n1.prepend(n1_2, &mut arena);
let n1_3 = arena.new_node("1_3");
n1.prepend(n1_3, &mut arena);

// arena
// `-- 1
//     |-- 1_3
//     |-- 1_2
//     `-- 1_1

let mut iter = n1.descendants(&arena);
assert_eq!(iter.next(), Some(n1));
assert_eq!(iter.next(), Some(n1_3));
assert_eq!(iter.next(), Some(n1_2));
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), None);

pub fn checked_prepend<T>(
self,
new_child: NodeId,
arena: &mut Arena<T>
) -> Result<(), NodeError>

Prepends a new child to this node, before existing children.

Failures

Returns NodeError::PrependSelf error if the given new child is self.

Returns NodeError::Removed error if the given new child or self is removed.

To check if the node is removed or not, use Node::is_removed().

Examples

let mut arena = Arena::new();
let n1 = arena.new_node("1");
assert!(n1.checked_prepend(n1, &mut arena).is_err());

let n1_1 = arena.new_node("1_1");
assert!(n1.checked_prepend(n1_1, &mut arena).is_ok());

pub fn insert_after<T>(self, new_sibling: NodeId, arena: &mut Arena<T>)

Inserts a new sibling after this node.

Panics

Panics if:

the given new sibling is self, or

the current node or the given new sibling was already removed.

To check if the node is removed or not, use Node::is_removed().

Examples

// arena
// `-- 1
//     |-- 1_1
//     `-- 1_2

let n1_3 = arena.new_node("1_3");
n1_1.insert_after(n1_3, &mut arena);

// arena
// `-- 1
//     |-- 1_1
//     |-- 1_3
//     `-- 1_2

let mut iter = n1.descendants(&arena);
assert_eq!(iter.next(), Some(n1));
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), Some(n1_3));
assert_eq!(iter.next(), Some(n1_2));
assert_eq!(iter.next(), None);

pub fn checked_insert_after<T>(
self,
new_sibling: NodeId,
arena: &mut Arena<T>
) -> Result<(), NodeError>

Inserts a new sibling after this node.

Failures

Returns NodeError::InsertAfterSelf error if the given new sibling is self.

Returns NodeError::Removed error if the given new sibling or self is removed.

To check if the node is removed or not, use Node::is_removed().

Examples

let mut arena = Arena::new();
let n1 = arena.new_node("1");
assert!(n1.checked_insert_after(n1, &mut arena).is_err());

let n2 = arena.new_node("2");
assert!(n1.checked_insert_after(n2, &mut arena).is_ok());

pub fn insert_before<T>(self, new_sibling: NodeId, arena: &mut Arena<T>)

Inserts a new sibling before this node.

Panics

Panics if:

the given new sibling is self, or

the current node or the given new sibling was already removed.

To check if the node is removed or not, use Node::is_removed().

Examples

let mut arena = Arena::new();
let n1 = arena.new_node("1");
let n1_1 = arena.new_node("1_1");
n1.append(n1_1, &mut arena);
let n1_2 = arena.new_node("1_2");
n1.append(n1_2, &mut arena);

// arena
// `-- 1
//     |-- 1_1
//     `-- 1_2

let n1_3 = arena.new_node("1_3");
n1_2.insert_before(n1_3, &mut arena);

// arena
// `-- 1
//     |-- 1_1
//     |-- 1_3
//     `-- 1_2

let mut iter = n1.descendants(&arena);
assert_eq!(iter.next(), Some(n1));
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), Some(n1_3));
assert_eq!(iter.next(), Some(n1_2));
assert_eq!(iter.next(), None);

pub fn checked_insert_before<T>(
self,
new_sibling: NodeId,
arena: &mut Arena<T>
) -> Result<(), NodeError>

Inserts a new sibling before this node.

Failures

Returns NodeError::InsertBeforeSelf error if the given new sibling is self.

Returns NodeError::Removed error if the given new sibling or self is removed.

To check if the node is removed or not, use Node::is_removed().

Examples

let mut arena = Arena::new();
let n1 = arena.new_node("1");
assert!(n1.checked_insert_before(n1, &mut arena).is_err());

let n2 = arena.new_node("2");
assert!(n1.checked_insert_before(n2, &mut arena).is_ok());

pub fn remove<T>(self, arena: &mut Arena<T>)

Removes a node from the arena.

Children of the removed node will be inserted to the place where the removed node was.

Please note that the node will not be removed from the internal arena storage, but marked as removed. Traversing the arena returns a plain iterator and contains removed elements too.

To check if the node is removed or not, use Node::is_removed().

Examples

// arena
// `-- 1
//     |-- 1_1
//     |-- 1_2
//     |   |-- 1_2_1
//     |   `-- 1_2_2
//     `-- 1_3

n1_2.remove(&mut arena);

let mut iter = n1.descendants(&arena);
assert_eq!(iter.next(), Some(n1));
assert_eq!(iter.next(), Some(n1_1));
assert_eq!(iter.next(), Some(n1_2_1));
assert_eq!(iter.next(), Some(n1_2_2));
assert_eq!(iter.next(), Some(n1_3));
assert_eq!(iter.next(), None);